Add create complaint validator unit tests
And reorganize all validators
