+++ src/AppServices/Staff/Validators/StaffUpdateValidator.cs
@@ -1,7 +1,7 @@
 ﻿using Cts.Domain.Identity;
 using FluentValidation;
 
-namespace Cts.AppServices.Staff;
+namespace Cts.AppServices.Staff.Validators;
 
 public class StaffUpdateValidator : AbstractValidator<StaffUpdateDto>
 {
