+++ tests/AppServicesTests/ActionTypes/Validators/UpdateValidator.cs
@@ -1,26 +1,27 @@
-﻿using Cts.AppServices.Offices;
-using Cts.Domain.Offices;
+﻿using Cts.AppServices.ActionTypes;
+using Cts.AppServices.ActionTypes.Validators;
+using Cts.Domain.ActionTypes;
 using Cts.TestData.Constants;
 using FluentValidation.TestHelper;
 
-namespace AppServicesTests.Offices;
+namespace AppServicesTests.ActionTypes;
 
 public class UpdateValidator
 {
     [Test]
     public async Task ValidDto_ReturnsAsValid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IActionTypeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Office?)null);
-        var model = new OfficeUpdateDto
+            .ReturnsAsync((ActionType?)null);
+        var model = new ActionTypeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new OfficeUpdateValidator(repoMock.Object);
+        var validator = new ActionTypeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -29,17 +30,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IActionTypeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new Office(Guid.NewGuid(), TestConstants.ValidName));
-        var model = new OfficeUpdateDto
+            .ReturnsAsync(new ActionType(Guid.NewGuid(), TestConstants.ValidName));
+        var model = new ActionTypeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new OfficeUpdateValidator(repoMock.Object);
+        var validator = new ActionTypeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name)
@@ -49,17 +50,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ForSameId_ReturnsAsValid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IActionTypeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new Office(Guid.Empty, TestConstants.ValidName));
-        var model = new OfficeUpdateDto
+            .ReturnsAsync(new ActionType(Guid.Empty, TestConstants.ValidName));
+        var model = new ActionTypeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new OfficeUpdateValidator(repoMock.Object);
+        var validator = new ActionTypeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -68,12 +69,12 @@ public class UpdateValidator
     [Test]
     public async Task NameTooShort_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IActionTypeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Office?)null);
-        var model = new OfficeUpdateDto() { Name = TestConstants.ShortName };
+            .ReturnsAsync((ActionType?)null);
+        var model = new ActionTypeUpdateDto() { Name = TestConstants.ShortName };
 
-        var validator = new OfficeUpdateValidator(repoMock.Object);
+        var validator = new ActionTypeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name);
