+++ tests/AppServicesTests/Concerns/Validators/CreateValidator.cs
@@ -1,21 +1,22 @@
-﻿using Cts.AppServices.Offices;
-using Cts.Domain.Offices;
+﻿using Cts.AppServices.Concerns;
+using Cts.AppServices.Concerns.Validators;
+using Cts.Domain.Concerns;
 using Cts.TestData.Constants;
 using FluentValidation.TestHelper;
 
-namespace AppServicesTests.Offices;
+namespace AppServicesTests.Concerns;
 
 public class CreateValidator
 {
     [Test]
     public async Task ValidDto_ReturnsAsValid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Office?)null);
-        var model = new OfficeCreateDto { Name = TestConstants.ValidName };
+            .ReturnsAsync((Concern?)null);
+        var model = new ConcernCreateDto { Name = TestConstants.ValidName };
 
-        var validator = new OfficeCreateValidator(repoMock.Object);
+        var validator = new ConcernCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -24,12 +25,12 @@ public class CreateValidator
     [Test]
     public async Task DuplicateName_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new Office(Guid.Empty, TestConstants.ValidName));
-        var model = new OfficeCreateDto { Name = TestConstants.ValidName };
+            .ReturnsAsync(new Concern(Guid.Empty, TestConstants.ValidName));
+        var model = new ConcernCreateDto { Name = TestConstants.ValidName };
 
-        var validator = new OfficeCreateValidator(repoMock.Object);
+        var validator = new ConcernCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name)
@@ -39,12 +40,12 @@ public class CreateValidator
     [Test]
     public async Task NameTooShort_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IOfficeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Office?)null);
-        var model = new OfficeCreateDto { Name = TestConstants.ShortName };
+            .ReturnsAsync((Concern?)null);
+        var model = new ConcernCreateDto { Name = TestConstants.ShortName };
 
-        var validator = new OfficeCreateValidator(repoMock.Object);
+        var validator = new ConcernCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name);
