+++ tests/AppServicesTests/Concerns/Validators/UpdateValidator.cs
@@ -1,26 +1,27 @@
-﻿using Cts.AppServices.ActionTypes;
-using Cts.Domain.ActionTypes;
+﻿using Cts.AppServices.Concerns;
+using Cts.AppServices.Concerns.Validators;
+using Cts.Domain.Concerns;
 using Cts.TestData.Constants;
 using FluentValidation.TestHelper;
 
-namespace AppServicesTests.ActionTypes;
+namespace AppServicesTests.Concerns;
 
 public class UpdateValidator
 {
     [Test]
     public async Task ValidDto_ReturnsAsValid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((ActionType?)null);
-        var model = new ActionTypeUpdateDto
+            .ReturnsAsync((Concern?)null);
+        var model = new ConcernUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ActionTypeUpdateValidator(repoMock.Object);
+        var validator = new ConcernUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -29,17 +30,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new ActionType(Guid.NewGuid(), TestConstants.ValidName));
-        var model = new ActionTypeUpdateDto
+            .ReturnsAsync(new Concern(Guid.NewGuid(), TestConstants.ValidName));
+        var model = new ConcernUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ActionTypeUpdateValidator(repoMock.Object);
+        var validator = new ConcernUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name)
@@ -49,17 +50,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ForSameId_ReturnsAsValid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new ActionType(Guid.Empty, TestConstants.ValidName));
-        var model = new ActionTypeUpdateDto
+            .ReturnsAsync(new Concern(Guid.Empty, TestConstants.ValidName));
+        var model = new ConcernUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ActionTypeUpdateValidator(repoMock.Object);
+        var validator = new ConcernUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -68,12 +69,12 @@ public class UpdateValidator
     [Test]
     public async Task NameTooShort_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IConcernRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((ActionType?)null);
-        var model = new ActionTypeUpdateDto() { Name = TestConstants.ShortName };
+            .ReturnsAsync((Concern?)null);
+        var model = new ConcernUpdateDto() { Name = TestConstants.ShortName };
 
-        var validator = new ActionTypeUpdateValidator(repoMock.Object);
+        var validator = new ConcernUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name);
