+++ tests/AppServicesTests/Offices/Validators/CreateValidator.cs
@@ -1,21 +1,22 @@
-﻿using Cts.AppServices.ActionTypes;
-using Cts.Domain.ActionTypes;
+﻿using Cts.AppServices.Offices;
+using Cts.AppServices.Offices.Validators;
+using Cts.Domain.Offices;
 using Cts.TestData.Constants;
 using FluentValidation.TestHelper;
 
-namespace AppServicesTests.ActionTypes;
+namespace AppServicesTests.Offices;
 
 public class CreateValidator
 {
     [Test]
     public async Task ValidDto_ReturnsAsValid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((ActionType?)null);
-        var model = new ActionTypeCreateDto { Name = TestConstants.ValidName };
+            .ReturnsAsync((Office?)null);
+        var model = new OfficeCreateDto { Name = TestConstants.ValidName };
 
-        var validator = new ActionTypeCreateValidator(repoMock.Object);
+        var validator = new OfficeCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -24,12 +25,12 @@ public class CreateValidator
     [Test]
     public async Task DuplicateName_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new ActionType(Guid.Empty, TestConstants.ValidName));
-        var model = new ActionTypeCreateDto { Name = TestConstants.ValidName };
+            .ReturnsAsync(new Office(Guid.Empty, TestConstants.ValidName));
+        var model = new OfficeCreateDto { Name = TestConstants.ValidName };
 
-        var validator = new ActionTypeCreateValidator(repoMock.Object);
+        var validator = new OfficeCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name)
@@ -39,12 +40,12 @@ public class CreateValidator
     [Test]
     public async Task NameTooShort_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IActionTypeRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((ActionType?)null);
-        var model = new ActionTypeCreateDto { Name = TestConstants.ShortName };
+            .ReturnsAsync((Office?)null);
+        var model = new OfficeCreateDto { Name = TestConstants.ShortName };
 
-        var validator = new ActionTypeCreateValidator(repoMock.Object);
+        var validator = new OfficeCreateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name);
