+++ tests/AppServicesTests/Offices/Validators/UpdateValidator.cs
@@ -1,26 +1,27 @@
-﻿using Cts.AppServices.Concerns;
-using Cts.Domain.Concerns;
+﻿using Cts.AppServices.Offices;
+using Cts.AppServices.Offices.Validators;
+using Cts.Domain.Offices;
 using Cts.TestData.Constants;
 using FluentValidation.TestHelper;
 
-namespace AppServicesTests.Concerns;
+namespace AppServicesTests.Offices;
 
 public class UpdateValidator
 {
     [Test]
     public async Task ValidDto_ReturnsAsValid()
     {
-        var repoMock = new Mock<IConcernRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Concern?)null);
-        var model = new ConcernUpdateDto
+            .ReturnsAsync((Office?)null);
+        var model = new OfficeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ConcernUpdateValidator(repoMock.Object);
+        var validator = new OfficeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -29,17 +30,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IConcernRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new Concern(Guid.NewGuid(), TestConstants.ValidName));
-        var model = new ConcernUpdateDto
+            .ReturnsAsync(new Office(Guid.NewGuid(), TestConstants.ValidName));
+        var model = new OfficeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ConcernUpdateValidator(repoMock.Object);
+        var validator = new OfficeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name)
@@ -49,17 +50,17 @@ public class UpdateValidator
     [Test]
     public async Task DuplicateName_ForSameId_ReturnsAsValid()
     {
-        var repoMock = new Mock<IConcernRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync(new Concern(Guid.Empty, TestConstants.ValidName));
-        var model = new ConcernUpdateDto
+            .ReturnsAsync(new Office(Guid.Empty, TestConstants.ValidName));
+        var model = new OfficeUpdateDto
         {
             Id = Guid.Empty,
             Name = TestConstants.ValidName,
             Active = true,
         };
 
-        var validator = new ConcernUpdateValidator(repoMock.Object);
+        var validator = new OfficeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldNotHaveValidationErrorFor(e => e.Name);
@@ -68,12 +69,12 @@ public class UpdateValidator
     [Test]
     public async Task NameTooShort_ReturnsAsInvalid()
     {
-        var repoMock = new Mock<IConcernRepository>();
+        var repoMock = new Mock<IOfficeRepository>();
         repoMock.Setup(l => l.FindByNameAsync(It.IsAny<string>(), It.IsAny<CancellationToken>()))
-            .ReturnsAsync((Concern?)null);
-        var model = new ConcernUpdateDto() { Name = TestConstants.ShortName };
+            .ReturnsAsync((Office?)null);
+        var model = new OfficeUpdateDto() { Name = TestConstants.ShortName };
 
-        var validator = new ConcernUpdateValidator(repoMock.Object);
+        var validator = new OfficeUpdateValidator(repoMock.Object);
         var result = await validator.TestValidateAsync(model);
 
         result.ShouldHaveValidationErrorFor(e => e.Name);
